+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ## Enhancements:
 
 - `starting-project` input added
+- env variables treated from commands
+- export command treated
 
 ## Bugs: